requirements.txt
streamlit>= 1.43.2
pandas>= 2.2.3
numpy>= 2.2.2
plotly>= 6.0.1
langgraph>= 0.3.23
langchain>= 0.3.21
pydantic>= 2.11.0
langchain-community>= 0.3.19
langchain-core>= 0.3.49
langchain-openai>= 0.3.10
python-dotenv>= 1.1.0
fastapi[standard]
python-jose[cryptography]
uvicorn[standard]
passlib[bcrypt]
slowapi
requests>=2.31.0
